make well formed its own file
