+++ edgelet/iotedge/src/check/checks/mod.rs
@@ -1,0 +1,2 @@
+pub mod well_formed_config;
+pub use self::well_formed_config::WellFormedConfig;
